# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Book&Bite",
    "version": "1.0.0",
    "description": "Documentación de la API del restaurante con Express y PostgreSQL"
  },
  "paths": {
    "/productos": {
      "get": {
        "summary": "Obtener todos los productos",
        "tags": [
          "Productos"
        ],
        "responses": {
          "200": {
            "description": "Lista de productos"
          },
          "500": {
            "description": "Error al obtener los productos"
          }
        }
      }
    },
    "/pedidos": {
      "post": {
        "summary": "Crear nuevo pedido",
        "tags": [
          "Pedidos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario_id": {
                    "type": "integer"
                  },
                  "total": {
                    "type": "number"
                  },
                  "estado": {
                    "type": "string"
                  },
                  "fecha": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido creado con éxito"
          },
          "500": {
            "description": "Error al crear el pedido"
          }
        }
      }
    },
    "/productos/alergenos": {
      "get": {
        "summary": "Obtener productos con sus alérgenos",
        "tags": [
          "Productos"
        ],
        "responses": {
          "200": {
            "description": "Lista de productos con sus alérgenos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "producto_id": {
                        "type": "integer"
                      },
                      "producto_nombre": {
                        "type": "string"
                      },
                      "alergenos": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener productos"
          }
        }
      }
    },
    "/reservas": {
      "post": {
        "summary": "Crear una nueva reserva",
        "tags": [
          "Reservas"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "usuario_id": {
                    "type": "integer"
                  },
                  "mesa_id": {
                    "type": "integer"
                  },
                  "fecha_hora": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "num_comensales": {
                    "type": "integer"
                  },
                  "estado": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reserva creada con éxito"
          },
          "500": {
            "description": "Error al crear la reserva"
          }
        }
      },
      "get": {
        "summary": "Obtener reservas futuras o filtrar por fecha",
        "tags": [
          "Reservas"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "fecha",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Fecha específica para filtrar reservas (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de reservas"
          },
          "404": {
            "description": "Reserva no encontrada"
          },
          "500": {
            "description": "Error al obtener las reservas"
          }
        }
      }
    },
    "/admin/pedidos": {
      "get": {
        "summary": "Obtener todos los pedidos (solo admin)",
        "tags": [
          "Pedidos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos"
          },
          "500": {
            "description": "Error al obtener los pedidos"
          }
        }
      }
    },
    "/admin/reservas": {
      "get": {
        "summary": "Obtener todas las reservas (solo admin)",
        "tags": [
          "Reservas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de reservas"
          },
          "500": {
            "description": "Error al obtener las reservas"
          }
        }
      }
    },
    "/productos/{id}": {
      "get": {
        "summary": "Obtener un producto por ID (incluye alérgenos)",
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del producto"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del producto"
          },
          "404": {
            "description": "Producto no encontrado"
          },
          "500": {
            "description": "Error al obtener el producto"
          }
        }
      }
    },
    "/usuarios": {
      "post": {
        "summary": "Crear un nuevo usuario",
        "tags": [
          "Usuarios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre",
                  "apellidos",
                  "email",
                  "telefono",
                  "direccion",
                  "contraseña"
                ],
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "apellidos": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "telefono": {
                    "type": "string"
                  },
                  "direccion": {
                    "type": "string"
                  },
                  "contraseña": {
                    "type": "string"
                  },
                  "rol": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Usuario creado con éxito"
            },
            "400": {
              "description": "Faltan campos requeridos"
            },
            "500": {
              "description": "Error al crear usuario"
            }
          }
        }
      }
    },
    "/usuarios/{id}/pedidos": {
      "get": {
        "summary": "Obtener pedidos de un usuario",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos"
          },
          "500": {
            "description": "Error al obtener los pedidos"
          }
        }
      }
    },
    "/reservas/{id}": {
      "put": {
        "summary": "Modificar estado o fecha de una reserva",
        "tags": [
          "Reservas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID de la reserva a modificar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fecha": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "estado": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reserva actualizada"
          },
          "404": {
            "description": "Reserva no encontrada"
          },
          "500": {
            "description": "Error al modificar reserva"
          }
        }
      }
    },
    "/alergenos": {
      "get": {
        "summary": "Listar todos los alérgenos",
        "tags": [
          "Productos"
        ],
        "responses": {
          "200": {
            "description": "Lista de alérgenos"
          },
          "500": {
            "description": "Error al obtener alérgenos"
          }
        }
      }
    },
    "/productos/{id}/alergenos": {
      "get": {
        "summary": "Obtener alérgenos de un producto",
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del producto"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de alérgenos del producto"
          },
          "404": {
            "description": "Producto no encontrado o sin alérgenos"
          },
          "500": {
            "description": "Error al obtener alérgenos del producto"
          }
        }
      }
    },
    "/valoraciones/{producto_id}": {
      "get": {
        "summary": "Obtener valoraciones de un producto",
        "tags": [
          "Productos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "producto_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del producto"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de valoraciones"
          },
          "500": {
            "description": "Error al obtener valoraciones"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Registrar nuevo usuario",
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "apellidos": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "contraseña": {
                    "type": "string"
                  }
                },
                "required": [
                  "nombre",
                  "apellidos",
                  "email",
                  "contraseña"
                ]
              }
            }
          },
          "responses": {
            "201": {
              "description": "Registrado con éxito"
            },
            "400": {
              "description": "Faltan campos requeridos"
            },
            "401": {
              "description": "Ya existe un usuario con ese email"
            },
            "500": {
              "description": "Error en registro"
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Iniciar sesión",
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre_usuario": {
                    "type": "string"
                  },
                  "contraseña": {
                    "type": "string"
                  }
                },
                "required": [
                  "nombre_usuario",
                  "contraseña"
                ]
              }
            }
          },
          "responses": {
            "200": {
              "description": "Inicio de sesión exitoso"
            },
            "400": {
              "description": "Nombre de usuario y contraseña requeridos"
            },
            "401": {
              "description": "Contraseña no válida o usuario no encontrado"
            },
            "500": {
              "description": "Error en el servidor"
            }
          }
        }
      }
    },
    "/usuarios/{id}": {
      "delete": {
        "summary": "Eliminar un usuario por ID (solo administradores)",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID del usuario a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error al eliminar usuario"
          }
        }
      }
    },
    "/eliminar-cuenta": {
      "delete": {
        "summary": "Eliminar tu propia cuenta",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tu cuenta ha sido eliminada"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error al eliminar tu cuenta"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}